divide composable
