+++ src/pages/composables/useNewRequest.ts
@@ -1,14 +1,14 @@
-import type { AxiosResponse } from 'axios'
 import { ref } from 'vue'
 import { useRouter } from 'vue-router'
 import { useToast } from 'vue-toastification'
 
 import { useRequestStore } from '/@/stores/request'
-import { useTagStore } from '/@/stores/tag'
 import { useUserStore } from '/@/stores/user'
 
-import type { Request, Tag } from '/@/lib/apis'
+import type { Request } from '/@/lib/apis'
 import apis from '/@/lib/apis'
+
+import { usePostTags } from './usePostTags'
 
 export interface FileRequest {
   name: string
@@ -28,9 +28,9 @@ export interface RequestRequest {
 export const useNewRequest = () => {
   const toast = useToast()
   const router = useRouter()
-  const tagStore = useTagStore()
   const requestStore = useRequestStore()
   const userStore = useUserStore()
+  const { postTags } = usePostTags()
 
   const request = ref<RequestRequest>({
     created_by: userStore.me?.name ?? '',
@@ -56,30 +56,12 @@ export const useNewRequest = () => {
       toast.warning('タイトル、内容、対象者は必須です')
       return
     }
-    const postTagResults: ReturnType<typeof apis.postTag>[] = []
-    let tags: string[] = []
-    // 以下の2パターンで場合分けしていて、条件式が1でtrue、2でfalseになる
-    // 1. willPutRequest.tagsに入っているidがtagStore.tagsのタグのidと一致する
-    // 2. タグが新規作成されて、そのタグ名がwillPutRequest.tagsに入っている
-    // 2の場合にはタグ名がtagStore.tagsのタグidと一致することはないので、tagPostPromisesにpostTag関数がpushされる
-    request.value.tags.forEach((tag: string) => {
-      if (tagStore.tags?.some(t => t.id === tag)) {
-        tags.push(tag)
-        return
-      }
-      postTagResults.push(apis.postTag({ name: tag }))
-    })
+    let tags
     try {
-      tags = tags.concat(
-        (await Promise.all(postTagResults)).map(
-          (tag: AxiosResponse<Tag>) => tag.data.id
-        )
-      )
+      tags = await postTags(request.value.tags)
     } catch {
-      toast.error('申請の作成に失敗しました')
       return
     }
-
     const requestRequest = {
       ...request.value,
       tags: tags,
