+++ src/pages/composables/useRequestDetail.ts
@@ -1,13 +1,12 @@
-import type { AxiosResponse } from 'axios'
 import { storeToRefs } from 'pinia'
 import { ref } from 'vue'
 import { useToast } from 'vue-toastification'
 
 import { useRequestDetailStore } from '/@/stores/requestDetail'
-import { useTagStore } from '/@/stores/tag'
 
-import type { Tag } from '/@/lib/apis'
 import apis from '/@/lib/apis'
+
+import { usePostTags } from './usePostTags'
 
 export type EditMode =
   | 'title'
@@ -30,8 +29,8 @@ export interface EditedValue {
 
 export const useRequestDetail = () => {
   const requestDetailStore = useRequestDetailStore()
-  const tagStore = useTagStore()
   const toast = useToast()
+  const { postTags } = usePostTags()
   const { request } = storeToRefs(requestDetailStore)
 
   const editMode = ref<EditMode>('')
@@ -54,28 +53,12 @@ export const useRequestDetail = () => {
     }
     editMode.value = ''
   }
+
   const putRequest = async (id: string, willPutRequest: EditedValue) => {
-    const tagPostPromises: ReturnType<typeof apis.postTag>[] = []
-    let tags: string[] = []
-    // 以下の2パターンで場合分けしていて、条件式が1でtrue、2でfalseになる
-    // 1. willPutRequest.tagsに入っているidがtagStore.tagsのタグのidと一致する
-    // 2. タグが新規作成されて、そのタグ名がwillPutRequest.tagsに入っている
-    // 2の場合にはタグ名がtagStore.tagsのタグidと一致することはないので、tagPostPromisesにpostTag関数がpushされる
-    willPutRequest.tags.forEach((tag: string) => {
-      if (tagStore.tags?.some(t => t.id === tag)) {
-        tags.push(tag)
-        return
-      }
-      tagPostPromises.push(apis.postTag({ name: tag }))
-    })
+    let tags
     try {
-      tags = tags.concat(
-        (await Promise.all(tagPostPromises)).map(
-          (tag: AxiosResponse<Tag>) => tag.data.id
-        )
-      )
+      tags = await postTags(willPutRequest.tags)
     } catch {
-      toast.error('申請の修正に失敗しました')
       return
     }
     const putRequest = {
